fixed namespace rename
