+++ src/Detection.Crawler/test/CrawlerTest.cs
@@ -4,7 +4,7 @@
 using Microsoft.AspNetCore.Http;
 
 using Moq;
-
+using Wangkanai.Detection.Services;
 using Xunit;
 
 namespace Wangkanai.Detection.Test
