+++ src/Detection/test/DetectionCoreollectionExtensionsTests.cs
@@ -6,7 +6,7 @@ using System.Collections.Generic;
 
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.DependencyInjection;
-
+using Wangkanai.Detection.Services;
 using Xunit;
 
 namespace Wangkanai.Detection.Test
